+++ src/string.js
@@ -1,6 +1,24 @@
 /**
- * Returns a hash from a given string
- * @param {String} str
+ * Truncates a given string to a given max length and appends a given string at
+ * the end.
+ * @param {string} string String to be truncated
+ * @param {number} maxLength Max length of the truncated string
+ * @param {string} append String to be appended to the truncated string
+ */
+const wordbreakTruncate = (string, maxLength = 140, append = '') => {
+  if (maxLength >= string.length) {
+    return string
+  }
+  let truncated = string.substr(0, maxLength + 1)
+  const end = truncated.lastIndexOf(' ')
+  truncated = truncated.substr(0, end)
+  const appendix = truncated.length < string.length ? append : ''
+  return truncated + appendix
+}
+
+/**
+ * Returns a hash from a given string.
+ * @param {string} str
  */
 const strToHash = str => {
   let hash = 0,
@@ -15,5 +33,6 @@ const strToHash = str => {
 }
 
 module.exports = {
+  wordbreakTruncate,
   strToHash,
 }